update pom.xml and versions

+++ pro/src/main/java/com/pro_servises/pro/model/Rating.java
@@ -16,7 +16,7 @@ public class Rating {
 
     @Id
     @GeneratedValue( strategy = GenerationType.AUTO)
-    public Long ratingId;
+    public Integer ratingId;
     private int stars;
     private String comment;
