+++ pro/src/main/java/com/pro_servises/pro/service/ProductService.java
@@ -8,9 +8,10 @@ public interface ProductService {
 
     public Product addProduct(Product product, Integer provider_id);
 
+
     Product findProductById(Long productId);
 
-    public List<Product> findAllProductsByProviderId(Long provider_id);
+    public List<Product> findAllProductsByProviderId(Integer provider_id);
 
     public void deleteProduct(Long productId);
 }
